Move Broadcaster accept logic to acceptor
This is a step towards factoring the message specific logic out of the
low-level broadcaster code.
